requirements.txt
itemadapter~=0.7.0
itemloaders~=1.0.6
PyDispatcher~=2.0.6
python-dotenv~=0.21.0
Scrapy~=2.6.2
w3lib~=2.0.1
pymongo~=4.2.0
python-dateutil~=2.8.2
typing~=3.7.4.3
tqdm~=4.64.1
spacy~=3.4.1
fuzzywuzzy~=0.18.0
requests~=2.28.1
bs4~=0.0.1
beautifulsoup4~=4.11.1
flair~=0.11.3
pandas~=1.4.4